done: find Windows API declaration

+++ Libraries/scintilla/src/Document.cxx
@@ -471,9 +471,9 @@ bool SCI_METHOD Document::Sci_SetFoldLevels(int line, int lineTo, int len, int* 
 	auto levels = Levels();
 	int levelCurrent = line == 0 ? SC_FOLDLEVELBASE : levels->GetLevel((Sci::Line)line - 1) >> 16;
 	int levelNext = levelCurrent;
-	int nLines = LinesTotal();
+	int nLines = (int)LinesTotal();
 	for(int i = 0; line < lineTo; line++) {
-		int eol = this->LineStart((Sci::Line)line + 1);
+		int eol = (int)this->LineStart((Sci::Line)line + 1);
 		if(a != nullptr) for(; i < len && (a[i] & 0x7fffffff) < eol; i++) levelNext += (a[i] & 0x80000000) ? -1 : 1;
 		if(levelNext < SC_FOLDLEVELBASE)levelNext = SC_FOLDLEVELBASE;
 		int lev = levelCurrent | levelNext << 16;
